proc: Use openat() for reading procfs when possible
This prevents a rare, but possible race condition where a process
ID can be killed and re-used while we're in the middle of reading
from its procfs entry.  Using openat() to read from the procfs
entry will correctly error out when the process is killed, even if
its process ID is re-used.

We still support the old method using open() on very old kernels
where openat() is not supported, or if we happen to run out of
file descriptors.

Signed-off-by: Jason Francis <zxzax@protonmail.com>