pgrep: Use pidfd_send_signal() when possible
This prevents another rare race condition where a process ID could
be killed and re-used before we get the chance to send a signal to
it, thus accidentally killing the wrong process.  We still fall
back to the old kill/sigqueue method on older kernels where
pidfd_send_signal() is not supported.

Signed-off-by: Jason Francis <zxzax@protonmail.com>