po-man: Install the corresponding po-man according to the configure option
We encountered a problem on the OpenCloudOS operating system:
Error shown:
  file /usr/share/man/de/man1/kill.1.gz from install of util-linux conflicts with file from package procps-ng
  file /usr/share/man/fr/man1/kill.1.gz from install of util-linux conflicts with file from package procps-ng
  file /usr/share/man/uk/man1/kill.1.gz from install of util-linux conflicts with file from package procps-ng

We prefer to obtain the kill command from util-linux like redhat, so
we disable procps' kill by configure --disable-kill. However, because
procps still provides the kill.1 po-man, there is a conflict with util-linux.

If we configure with -disable-kill, we don't need any kill.1 man either,
this is more reasonable.

The po-man we want to install is defined according to configure options.

Signed-off-by: Chunsheng Luo <luochunsheng@ustc.edu>