Makefile.am: Add `check-man` target
Add target to run groff(1) with all warnings enabled, and set the
`CHECKSTYLE` register, recognized by the man(7) in groff 1.23.0 (and
silently ignored elsewhere).  Output is discarded.

Don't attempt to check man/{pkill,pidwait}.1, because they are stub man
pages consisting only of a *roff `so` request using a path that is
resolvable only once installed.