sysctl: Ensure nested errors affect the exit code
When reading nested sysctls, such as with the -a option or by specifying
a directory (e.g. net.ipv6), sysctl would return an exit code of 0 even
if it encountered errors while reading individual keys. This made it
impossible for scripts to reliably detect when a read operation was only
partially successful.

This change ensures that any failure is accumulated, and that a single
error during the traversal will correctly result in a non-zero final
exit code.

Before:

    $ sysctl net.ipv6 > /dev/null; echo $?
    sysctl: permission denied on key 'net.ipv6.conf.all.stable_secret'
    sysctl: permission denied on key 'net.ipv6.conf.default.stable_secret'
    sysctl: permission denied on key 'net.ipv6.conf.eth0.stable_secret'
    sysctl: permission denied on key 'net.ipv6.conf.lo.stable_secret'
    0

After:

    $ src/sysctl net.ipv6 > /dev/null; echo $?
    sysctl: permission denied on key 'net.ipv6.conf.all.stable_secret'
    sysctl: permission denied on key 'net.ipv6.conf.default.stable_secret'
    sysctl: permission denied on key 'net.ipv6.conf.eth0.stable_secret'
    sysctl: permission denied on key 'net.ipv6.conf.lo.stable_secret'
    1

References: #388

Signed-off-by: Chris Down <chris@chrisdown.name>